Napravljen mobile design za login/register/sidebar navigation
