+++ tv_shows_app/src/components/shared/SidebarNavigation/layouts/SidebarNavigation.desktop.tsx
@@ -1,0 +1,86 @@
+'use client';
+import theme from '@/styles/theme/theme';
+import { Flex, Text } from '@chakra-ui/react';
+import NextLink from 'next/link';
+import { useParams, usePathname, useRouter } from 'next/navigation';
+import { useEffect, useState } from 'react';
+import { useLocation } from 'react-router-dom';
+
+export const SidebarNavigationDesktop = () => {
+  const router = useRouter();
+  const pathname = usePathname();
+  const [path, setPath] = useState(pathname);
+  useEffect(() => {
+    setPath(pathname);
+  }, [pathname]);
+  return (
+    <Flex
+      width="100%"
+      flexDirection="column"
+      gap="10"
+      backgroundColor="darkPurple"
+      height="100%"
+      padding="4"
+    >
+      <Text color="white" fontWeight="700" textStyle="title">
+        TV shows APP
+      </Text>
+      <Flex flexDirection="column" color="white" gap="2">
+        <Text
+          textStyle="smallCaption"
+          pt="10px"
+          borderRadius="full"
+          width="140px"
+          height="44px"
+          textAlign="center"
+          as={NextLink}
+          href="/shows"
+          backgroundColor={path == '/shows' ? 'purple2' : 'darkPurple'}
+        >
+          All shows
+        </Text>
+        <Text
+          textStyle="smallCaption"
+          pt="10px"
+          borderRadius="full"
+          width="140px"
+          height="44px"
+          textAlign="center"
+          as={NextLink}
+          href="/top-rated"
+          backgroundColor={path == '/top-rated' ? 'purple2' : 'darkPurple'}
+        >
+          Top rated
+        </Text>
+        <Text
+          textStyle="smallCaption"
+          pt="10px"
+          borderRadius="full"
+          width="140px"
+          height="44px"
+          textAlign="center"
+          as={NextLink}
+          href="/my-profile"
+          backgroundColor={path == '/my-profile' ? 'purple2' : 'darkPurple'}
+        >
+          My profile
+        </Text>
+      </Flex>
+      <Text
+        width="90px"
+        marginTop="auto"
+        color="white"
+        as="button"
+        marginRight="auto"
+        _hover={{ color: 'red' }}
+        onClick={() => {
+          localStorage.removeItem('tv-shows-uid'),
+            localStorage.removeItem('tv-shows-header');
+          router.push('/login');
+        }}
+      >
+        Log out
+      </Text>
+    </Flex>
+  );
+};
