+++ tv_shows_app/src/components/shared/SidebarNavigation/layouts/SidebarNavigation.mobile.tsx
@@ -1,0 +1,134 @@
+import theme from '@/styles/theme/theme';
+import { HamburgerIcon } from '@chakra-ui/icons';
+import {
+  Button,
+  Drawer,
+  DrawerBody,
+  DrawerCloseButton,
+  DrawerContent,
+  DrawerFooter,
+  DrawerHeader,
+  DrawerOverlay,
+  Flex,
+  IconButton,
+  Input,
+  Text,
+  useDisclosure,
+} from '@chakra-ui/react';
+import NextLink from 'next/link';
+import { useParams, usePathname, useRouter } from 'next/navigation';
+import React from 'react';
+import { useEffect, useState } from 'react';
+import { useLocation } from 'react-router-dom';
+
+export const SidebarNavigationMobile = () => {
+  const router = useRouter();
+  const pathname = usePathname();
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  const [path, setPath] = useState(pathname);
+  useEffect(() => {
+    setPath(pathname);
+  }, [pathname]);
+  return (
+    <Flex
+      width="100%"
+      flexDirection="row"
+      gap="10"
+      backgroundColor="darkPurple"
+      height="100%"
+      padding="4"
+    >
+      <Text color="white" fontWeight="700" textStyle="title">
+        TV shows APP
+      </Text>
+      <>
+        <IconButton
+          ml="auto"
+          backgroundColor="transparent"
+          _hover={{ backgroundColor: 'transparent' }}
+          colorScheme="teal"
+          onClick={onOpen}
+          icon={<HamburgerIcon />}
+          size="md"
+          aria-label={'dropdown menu'}
+        >
+          Open
+        </IconButton>
+        <Drawer isOpen={isOpen} placement="right" onClose={onClose} size="xs">
+          <DrawerOverlay />
+          <DrawerContent backgroundColor="purple2">
+            <DrawerCloseButton color="white" border="1px" borderRadius="full" />
+
+            <DrawerBody mt="84px">
+              <Flex flexDirection="column" color="white" gap="2">
+                <Text
+                  textStyle="smallCaption"
+                  pt="10px"
+                  borderRadius="full"
+                  width="140px"
+                  height="44px"
+                  textAlign="center"
+                  as={NextLink}
+                  href="/shows"
+                  backgroundColor={
+                    path == '/shows' ? 'lightPurple' : 'transparent'
+                  }
+                >
+                  All shows
+                </Text>
+                <Text
+                  textStyle="smallCaption"
+                  pt="10px"
+                  borderRadius="full"
+                  width="140px"
+                  height="44px"
+                  textAlign="center"
+                  as={NextLink}
+                  href="/top-rated"
+                  backgroundColor={
+                    path == '/top-rated' ? 'lightPurple' : 'transparent'
+                  }
+                >
+                  Top rated
+                </Text>
+                <Text
+                  textStyle="smallCaption"
+                  pt="10px"
+                  borderRadius="full"
+                  width="140px"
+                  height="44px"
+                  textAlign="center"
+                  as={NextLink}
+                  href="/my-profile"
+                  backgroundColor={
+                    path == '/my-profile' ? 'lightPurple' : 'transparent'
+                  }
+                >
+                  My profile
+                </Text>
+              </Flex>
+            </DrawerBody>
+
+            <DrawerFooter>
+              <Text
+                width="90px"
+                marginTop="auto"
+                color="white"
+                as="button"
+                marginRight="auto"
+                _hover={{ color: 'red' }}
+                onClick={() => {
+                  localStorage.removeItem('tv-shows-uid'),
+                    localStorage.removeItem('tv-shows-header');
+                  router.push('/login');
+                }}
+              >
+                Log out
+              </Text>
+            </DrawerFooter>
+          </DrawerContent>
+        </Drawer>
+      </>
+    </Flex>
+  );
+};
