+++ tv_shows_app/src/components/shared/auth/PasswordInput.tsx
@@ -3,11 +3,13 @@ import { EmailIcon, LockIcon, ViewIcon, ViewOffIcon } from '@chakra-ui/icons';
 import {
   Button,
   FormControl,
+  Hide,
   IconButton,
   Input,
   InputGroup,
   InputLeftElement,
   InputRightElement,
+  Show,
   Text,
 } from '@chakra-ui/react';
 import { useState } from 'react';
@@ -23,37 +25,75 @@ interface IPasswordInput {
 export const PasswordInput = (params: IPasswordInput) => {
   const [hidden, setHidden] = useState(true);
   return (
-    <InputGroup>
-      <InputLeftElement
-        pointerEvents="none"
-        children={<LockIcon color="white" width="24px" />}
-        ml="56px"
-      />
-      <FormControl textColor="white">
-        <Input
-          isDisabled={params.disable}
-          {...params.thatPart}
-          required
-          placeholder={params.placeholder}
-          type={hidden ? 'password' : 'text'}
-          _placeholder={{ color: 'white' }}
-          size="md"
-          data-testid={params.testId}
-        />
-      </FormControl>
-      <InputRightElement fontSize="md">
-        <IconButton
-          marginRight="120px"
-          color="white"
-          aria-label="Search database"
-          icon={hidden ? <ViewIcon /> : <ViewOffIcon />}
-          onClick={() => {
-            setHidden(!hidden);
-          }}
-          backgroundColor="transparent"
-          _hover={{ backgroundColor: 'transparent' }}
-        />
-      </InputRightElement>
-    </InputGroup>
+    <>
+      <Show above="1024px">
+        <InputGroup>
+          <InputLeftElement
+            pointerEvents="none"
+            children={<LockIcon color="white" width="24px" />}
+            ml="56px"
+          />
+          <FormControl textColor="white">
+            <Input
+              isDisabled={params.disable}
+              {...params.thatPart}
+              required
+              placeholder={params.placeholder}
+              type={hidden ? 'password' : 'text'}
+              _placeholder={{ color: 'white' }}
+              size="md"
+              data-testid={params.testId}
+            />
+          </FormControl>
+          <InputRightElement fontSize="md">
+            <IconButton
+              marginRight="120px"
+              color="white"
+              aria-label="Search database"
+              icon={hidden ? <ViewIcon /> : <ViewOffIcon />}
+              onClick={() => {
+                setHidden(!hidden);
+              }}
+              backgroundColor="transparent"
+              _hover={{ backgroundColor: 'transparent' }}
+            />
+          </InputRightElement>
+        </InputGroup>
+      </Show>
+      <Hide above="1024px">
+        <InputGroup>
+          <InputLeftElement
+            pointerEvents="none"
+            children={<LockIcon color="white" width="24px" />}
+          />
+          <FormControl textColor="white">
+            <Input
+              mx="0"
+              isDisabled={params.disable}
+              {...params.thatPart}
+              required
+              placeholder={params.placeholder}
+              type={hidden ? 'password' : 'text'}
+              _placeholder={{ color: 'white' }}
+              size="sm"
+              data-testid={params.testId}
+            />
+          </FormControl>
+          <InputRightElement fontSize="md">
+            <IconButton
+              marginRight="0"
+              color="white"
+              aria-label="Search database"
+              icon={hidden ? <ViewIcon /> : <ViewOffIcon />}
+              onClick={() => {
+                setHidden(!hidden);
+              }}
+              backgroundColor="transparent"
+              _hover={{ backgroundColor: 'transparent' }}
+            />
+          </InputRightElement>
+        </InputGroup>
+      </Hide>
+    </>
   );
 };
